+++ src/ber.rs
@@ -531,7 +531,7 @@ impl Ber {
             buffer.set_len(total_len);
 
             let ptr = buffer.as_mut_ptr();
-            ptr.copy_from_nonoverlapping(id.as_ptr(), id.len());
+            ptr.copy_from_nonoverlapping(id.as_bytes().as_ptr(), id.len());
 
             let ptr = ptr.add(id.len());
             ptr.copy_from_nonoverlapping(length.as_ptr(), length.len());
@@ -613,7 +613,7 @@ impl Ber {
 
         unsafe {
             let dst = buffer.as_mut_ptr();
-            dst.copy_from_nonoverlapping(id.as_ptr(), id.len());
+            dst.copy_from_nonoverlapping(id.as_bytes().as_ptr(), id.len());
 
             let dst = dst.add(id.len());
             dst.copy_from_nonoverlapping(length_.as_ptr(), length_.len());