+++ CHANGELOG.md
@@ -158,4 +158,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 ### Removed
 - Delete the implementations for the following traits
     - Boffow<[u8]> for IdRef
+    - Deref for IdRef
     - DerefMut for IdRef